P8 fixed and guitar improved

+++ P8/p8-2.py
@@ -20,7 +20,7 @@ class Example(Base):
         print("Initializing program...")
         self.renderer = Renderer()
         self.scene = Scene()
-        self.camera = Camera(aspect_ratio=int(800/600))
+        self.camera = Camera(aspect_ratio=800/600)
         self.camera.set_position([0, 0, 5])
         geometry = Geometry()
         vs_code = """
